+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Add support for configuring a Context API
+- Add Rack middleware to store context in session
 
 ## [0.6.2, 0.5.2, 0.4.8] - 2025-07-21